+++ src/pyodide/internal/snapshot.ts
@@ -161,17 +161,6 @@ export function preloadDynamicLibs(Module: Module): void {
   }
 }
 
-export function getSnapshotSettings() {
-  return {
-    preRun: [preloadDynamicLibs],
-    // if SNAPSHOT_SIZE is defined, start with the linear memory big enough to
-    // fit the snapshot. If it's not defined, this falls back to the default.
-    INITIAL_MEMORY: SNAPSHOT_SIZE,
-    // skip running main() if we have a snapshot
-    noInitialRun: SHOULD_RESTORE_SNAPSHOT,
-  };
-}
-
 type DylinkInfo = {
   [name: string]: { handles: string[] };
 } & {
@@ -400,6 +389,7 @@ export function restoreSnapshot(Module: Module): void {
   if (!READ_MEMORY) {
     throw Error('READ_MEMORY not defined when restoring snapshot');
   }
+  Module.growMemory(SNAPSHOT_SIZE!);
   READ_MEMORY(Module);
 }
